+++ 2025/day-02/src/product.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <istream>
 
-int count_invalid_ids(std::istream &input);
+long count_invalid_ids(std::istream &input);
